+++ tests/test_router_bgp.rb
@@ -61,9 +61,11 @@ def setup_vrf
 end
 
 def newer_image_version?
-  return false if Utils.image_version?(/7.0.3.I2|I3|I4/) ||
-                  Utils.chassis_pid?(/N(5|6|7|8)/)
-  true
+  new = true
+  new = false if Utils.image_version?(/7.0.3.I2|I3|I4/) ||
+                 node.product_id[/(N5|N6|N7|N9.*-F)/]
+  new = true if Utils.image_version?(/8.0/)
+  new
 end
 
 # TestRouterBgp - Minitest for RouterBgp class
@@ -666,7 +668,8 @@ class TestRouterBgp < CiscoTestCase
     bgp.event_history_errors = 'false'
     assert_equal('false', bgp.event_history_errors)
     bgp.event_history_errors = 'size_small'
-    assert_equal('size_small', bgp.event_history_errors)
+    assert_equal('size_small', bgp.event_history_errors) unless
+      Utils.image_version?(/8.0/)
     bgp.event_history_errors = 'size_large'
     assert_equal('size_large', bgp.event_history_errors)
     bgp.event_history_errors = 'size_medium'
@@ -752,7 +755,8 @@ class TestRouterBgp < CiscoTestCase
     assert_equal(bgp.default_event_history_periodic,
                  bgp.event_history_periodic)
     bgp.event_history_periodic = 'false'
-    assert_equal('false', bgp.event_history_periodic)
+    assert_equal('false', bgp.event_history_periodic) unless
+      Utils.image_version?(/8.0/)
     bgp.event_history_periodic = 'size_small'
     assert_equal('size_small', bgp.event_history_periodic)
     bgp.event_history_periodic = 'size_large'
@@ -770,7 +774,8 @@ class TestRouterBgp < CiscoTestCase
     end
     bgp.event_history_periodic = 'true'
     if newer_image_version?
-      assert_equal('true', bgp.event_history_periodic)
+      assert_equal('true', bgp.event_history_periodic) unless
+        Utils.image_version?(/8.0/)
     else
       assert_equal(bgp.default_event_history_periodic,
                    bgp.event_history_periodic)